Index Changes, footer created and navbar modified
minor changes of index, footer is responsive and colocated in all sectiones,, and navbar has a button positionated on the right side which his function is show the page`s sections when someone click on his. All of new files has his own css file
I made changes on location section.
thanks for all
Kisses and hugs for everyone!